Split OnboardingWindow.py into Screens/ subfolder
